Membagi Class yang Besar
Membagi Class Employee yang besar menjadi beberapa Class yang lebih kecil
