+++ TaxCalculator/TaxCalculator/src/lib/Family.java
@@ -1,0 +1,23 @@
+package lib;
+import java.time.LocalDate;
+import java.time.Month;
+import java.util.LinkedList;
+import java.util.List;
+
+public class Family extends Person{
+    private String spouseName;
+	private String spouseIdNumber;
+
+	private List<String> childNames;
+	private List<String> childIdNumbers;
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = getIdNumber();
+	}
+
+    public void addChild(String childName, String childIdNumber) {
+		childNames.add(childName);
+		childIdNumbers.add(childIdNumber);
+	}
+}
